+++ tomcat/src/main/java/org/apache/coyote/request/startline/HttpRequestPath.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.request;
+package org.apache.coyote.request.startline;
 
 public class HttpRequestPath {
